# app/ios/openapi.yaml
openapi: 3.0.0
info:
  title: Public AR-Stolperstein API
  description: This is the Public OpenAPI Specification for the AR-Stolpersteine Backend. Every Endpoint is without an Authentication accessible
  version: 1.0.0
servers:
  - url: 'http://localhost:443'
paths:
  /api/public/stolpersteine:
    get:
      tags:
        - Public
      summary: Get Stolpersteine in a specific radius
      description: Get Stolpersteine in a specific radius (e.g. 500m)
      operationId: getStolpersteineForUser
      parameters:
        - name: radius
          in: query
          description: Area around current User-Location
          required: true
          schema:
            type: number
            format: double
            description: Number represent radius in meters
            example: 200
        - name: latUser
          in: query
          description: Lat-Coordinate from current User-Location
          required: true
          schema:
            type: number
            format: float
            example: 90.233232
        - name: lngUser
          in: query
          description: Lng-Coordinate from current User-Location
          required: true
          schema:
            type: number
            format: float
            example: 45.233232
      responses:
        200:
          description: return all stolpersteine based on current user location and radius
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StolpersteinPositionDto'
        400:
          description: Bad Request - All Query-Params need to be set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDto'
        404:
          description: If no Stolpersteine for UserId was found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDto'
        500:
          description: Internal Server Errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDto'
  /api/public/stolpersteine/{stolpersteinId}:
    get:
      tags:
        - Public
      summary: Get Stolpersteine for given stolpersteinId
      description: Get all available Stolpersteine for an User
      operationId: getStolpersteinForID
      parameters:
        - name: stolpersteinId
          in: path
          description: ID of Stolperstein
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: return a stolpersteine based for stolpersteinID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StolpersteineResponseDto'
        400:
          description: Bad Request - All Query-Params need to be set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDto'
        404:
          description: If no Stolpersteine for UserId was found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDto'
        500:
          description: Internal Server Errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDto'
  /api/public/photo/download/{resourceGroupId}/{resourceId}:
    get:
      tags:
        - Public
      summary: Download Photo for given ResourceURL
      description: Download Photo for given ResourceURL
      operationId: downloadPhoto
      parameters:
        - name: resourceGroupId
          in: path
          description: ID of Resource Group
          required: true
          schema:
            type: integer
            format: int64
        - name: resourceId
          in: path
          description: ID of Photo
          required: true
          schema:
            type: string
      responses:
        200:
          description: Return information when upload was successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        404:
          description: Resource was not found with given Resource-URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDto'
        400:
          description: Bad Request when no valid request was send
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDto'
        500:
          description: Internal Server Errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDto'
components:
  schemas:
    ProblemDto:
      type: object
      properties:
        errorCode:
          type: integer
          example: 400
        description:
          type: string
          example: Bad Request - Missing Parameter
    StolpersteinListResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 232
        position:
          type: object
          properties:
            lat:
              type: number
              format: float
              example: 90.233232
            lng:
              type: number
              format: float
              example: 23.232232
        victim:
          type: object
          properties:
            firstname:
              type: string
              example: Max
            lastname:
              type: string
              example: Mustermann
            dateOfDeath:
              type: string
              format: date
              example: 1942-02-23
    VictimResponseDto:
      type: object
      properties:
        firstname:
          type: string
          example: Max
        lastname:
          type: string
          example: Mustermann
        dateOfBirth:
          type: string
          format: date
          example: 1933-12-21
        dateOfDeath:
          type: string
          format: date
          example: 1942-02-23
    AddressResponseDto:
      type: object
      properties:
        streetName:
          type: string
          example: "Mustermannstrasse"
        houseNumber:
          type: string
          maxLength: 4
          minLength: 1
          example: 12a
        postCode:
          type: string
          maxLength: 5
          minLength: 5
          example: 97074
        city:
          type: string
          example: Wuerzburg
    PhotoResponseDto:
      type: object
      properties:
        heading:
          type: integer
          description: representing the heading on a compass for N, W, E, S
          format: int16
          example: 90
        resourceUrl:
          type: string
          description: specific path to resource to download the file
          example: "/232/example-crypt-3s2323.jpeg"
        resourceGroup:
          type: string
          description: representing the group where all available resources can be loaded as once
          example: "/232"
    StolpersteineResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1023
        description:
          type: string
          example: That is an example of a stolperstein description
        position:
          type: object
          properties:
            lat:
              type: number
              format: float
              example: 90.233232
            lng:
              type: number
              format: float
              example: 23.232232
        victim:
          $ref: '#/components/schemas/VictimResponseDto'
        address:
          $ref: '#/components/schemas/AddressResponseDto'
        photos:
          type: array
          items:
            $ref: '#/components/schemas/PhotoResponseDto'
    StolpersteinPositionDto:
      description: Object which represents only a single point of a stolperstein without any further informations
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1023
        position:
          type: object
          properties:
            lat:
              type: number
              format: float
              example: 90.233232
            lng:
              type: number
              format: float
              example: 23.232232